+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "anubrahman_backend"
+name = "axum_backend"
 version = "0.1.0"
 edition = "2024"
