+++ src/lib.rs
@@ -1,4 +1,4 @@
 pub mod dbs;
 pub mod err;
 pub use err::error::AppError;
-pub mod svr;
+pub mod sys;
